+++ BaAA/week14/e/set.cpp
@@ -1,10 +1,11 @@
 #include <ios>
 #include <iostream>
+#include <climits>
 #include <set>
 #include <vector>
 #include <utility>
 
-unsigned int getSumDigits(uint16_t number)
+unsigned int getSumDigits(unsigned int number)
 {
     unsigned int answer = 0;
     while (number > 0)
@@ -15,7 +16,7 @@ unsigned int getSumDigits(uint16_t number)
     return answer;
 }
 
-void printPath(const std::vector<uint16_t>& parents, uint16_t vertex, uint16_t dist = 0)
+void printPath(const std::vector<unsigned int>& parents, unsigned int vertex, unsigned int dist = 0)
 {
     if (parents[vertex] == 0) { std::cout << dist << '\n'; return; }
     printPath(parents, parents[vertex], dist + 1);
@@ -29,16 +30,16 @@ int main()
     std::ios::sync_with_stdio(false);
     std::cin.tie(nullptr);
 
-    uint16_t n, a, b;
+    unsigned int n, a, b;
     std::cin >> n >> a >> b;
     ++n;
 
     constexpr unsigned int INF = UINT_MAX;
-    std::vector<uint16_t> parents(n, 0);
+    std::vector<unsigned int> parents(n, 0);
     std::vector<unsigned int> distances(n, INF);
     distances[a] = 0;
 
-    std::set< std::pair<unsigned int, uint16_t> > to_update;
+    std::set< std::pair<unsigned int, unsigned int> > to_update;
     to_update.emplace(0, a);
     while (!to_update.empty())
     {
@@ -51,14 +52,14 @@ int main()
         }
         to_update.erase(to_update.begin());
 
-        uint16_t prefix = from;
+        unsigned int prefix = from;
         while (prefix > 0)
         {
-            uint16_t suffix = prefix;
+            unsigned int suffix = prefix;
             unsigned int power10 = 1;
             do
             {
-                uint16_t y = suffix % (power10 *= 10);
+                unsigned int y = suffix % (power10 *= 10);
                 unsigned int attempt_distance = current_distance + getSumDigits(y);
                 if (from + y < n && attempt_distance < distances[from + y])
                 {
